+++ library/src/main/java/com/wang/avi/indicators/BallSpinFadeLoaderIndicator.java
@@ -59,7 +59,7 @@ public class BallSpinFadeLoaderIndicator extends Indicator {
             ValueAnimator scaleAnim=ValueAnimator.ofFloat(1,0.4f,1);
             scaleAnim.setDuration(getDuration());
             scaleAnim.setRepeatCount(-1);
-            scaleAnim.setStartDelay(delays[i]*getMultiplier());
+            scaleAnim.setStartDelay((long) (delays[i]*getMultiplier()));
             addUpdateListener(scaleAnim,new ValueAnimator.AnimatorUpdateListener() {
                 @Override
                 public void onAnimationUpdate(ValueAnimator animation) {
@@ -71,7 +71,7 @@ public class BallSpinFadeLoaderIndicator extends Indicator {
             ValueAnimator alphaAnim=ValueAnimator.ofInt(255, 77, 255);
             alphaAnim.setDuration(getDuration());
             alphaAnim.setRepeatCount(-1);
-            alphaAnim.setStartDelay(delays[i]*getMultiplier());
+            alphaAnim.setStartDelay((long) (delays[i]*getMultiplier()));
             addUpdateListener(alphaAnim,new ValueAnimator.AnimatorUpdateListener() {
                 @Override
                 public void onAnimationUpdate(ValueAnimator animation) {
